Add initial formulas for Themis 0.10.0
Themis is actually a wrapper over low-level cryptographic libraries.
Currently it supports OpenSSL, LibreSSL, and BoringSSL as backends.
Homebrew offers only OpenSSL and LibreSSL so that's what we do.

The backend choice is mutually exclusive as the library is named just
"themis". The user needs to select a particular flavor.  We could have
provided a single formula with options (--with-{openssl,libressl}), but
such formulas are frowned upon in the core Homebrew repo, and I don't
want to check for mutual exclusion. So instead we're going with mutually
conflicting formulas. This allows the user to install both flavors
simultaneously, with the drawback of not having Themis in PATH anymore.