Add BoringSSL flavor
Themis is going to use BoringSSL as its main backend some day so let's
make it available via Homebrew. And make it the default flavor in order
to facilitate testing, ninja-style.

Please let me express my... condolences to all C++ users who have to
deal with this dependency management crap they have. BoringSSL makes it
especially visible.

First of all, it requires the following as build-time dependencies:

- CMake
- Golang
- Perl
- Make or Ninja
- C and C++ compilers

Some of these are reasonably available 'standard' UNIX tools, but why
the hell do I need Go? Because this is Google and it commands so.

Then comes the distribution and versioning policy. Except there is none
and you're supposed to grab whatever version you like and just use it,
there are no stability guarantees. Amazing. So I did just that: this
formula uses the current (as of 2018-11-01) snapshot of BoringSSL.

Next is the installation, which is missing. Apparently, install() macros
of CMake are too complex so I have to hardcode and copy the targets
manually.

And I'm not even talking about the fact that CMake ignores CFLAGS and
friends so I have to hardcode release-with-debug building mode in order
to get reasonable binaries (while the user and Homebrew may think
otherwise).

Well, whatever. This formula seems to work. It's all that matters to me.